Clear the pre-existing clippy warnings (0 warnings across all targets)
Six long-standing lib warnings plus three in tests, none behavior-affecting:

- link.rs / verify.rs: collapse a nested `if`/`if let` and a single-arm
  `match` into edition-2024 let-chains (the idiom already used widely here).
- op_registry.rs, trace.rs, types.rs (type_expr_with_locals, type_builtin_box_new):
  annotate the four context-threading functions with
  `#[allow(clippy::too_many_arguments)]`, consistent with the ~50 existing uses.
- tests/pattern_match_native.rs: replace `100 + 1 + 2 + 0 + 9 + 0` (identity_op on
  the `+ 0` terms) with the literal `112` plus a comment mapping each classify case.
- tests/oracle_conformance.rs: drop a useless `u32::from` on an already-`u32` value.

Pure cleanup: full suite green (502 tests, 0 failures); `cargo clippy --all-targets`
now reports zero warnings.

+++ tests/oracle_conformance.rs
@@ -80,7 +80,7 @@ const WIDTHS: &[(&str, u32, bool)] = &[
 /// the decimal text the `--expect-int` flag wants — the wrapping the evaluator and
 /// backend both apply.
 fn wrap_to_width(value: i128, width_bytes: u32, signed: bool) -> String {
-    let bits = u32::from(width_bytes) * 8;
+    let bits = width_bytes * 8;
     let modulus = 1i128 << bits;
     let masked = value.rem_euclid(modulus);
     if signed && masked >= (1i128 << (bits - 1)) {
